backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.35
pydantic==2.8.2
python-dotenv==1.0.1

